+++ src/Raycynix.Extensions.Database.MsSql/Configurations/MsSqlServerConfiguration.cs
@@ -8,7 +8,7 @@ public class MsSqlServerConfiguration
     /// <summary>
     /// Gets a value indicating whether the server certificate should be trusted.
     /// </summary>
-    public bool TrustServerCertificate { get; set; } = true;
+    public bool TrustServerCertificate { get; set; } = false;
 
     /// <summary>
     /// Gets the command timeout in seconds.