+++ src/Raycynix.Extensions.Database.Hosting/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.0.0
+### Added
+- Added generic host and service provider startup helpers for running `IDatabaseInitializer`.
+
 ## 1.0.2
 ### Added
 - Initial package release.